+++ src/silicone/database_crunchers/__init__.py
@@ -12,6 +12,7 @@ from .equal_quantile_walk import EqualQuantileWalk  # noqa: F401
 from .interpolate_specified_scenarios_and_models import (  # noqa: F401
     ScenarioAndModelSpecificInterpolate,
 )
+from .interpolation import Interpolation  # noqa: F401
 from .latest_time_ratio import LatestTimeRatio  # noqa: F401
 from .linear_interpolation import LinearInterpolation  # noqa: F401
 from .quantile_rolling_windows import QuantileRollingWindows  # noqa: F401